Move the metadata parsing (for display) to separate service. Update parsing logic to exclude 'issn-map'
