+++ app/components/workflow-review.js
@@ -1,13 +1,15 @@
 import Component from '@ember/component';
-import _ from 'lodash';
 
 export default Component.extend({
+  metatadaService: Ember.inject.service('metadata-blob'),
+
   init() {
     this._super(...arguments);
     // // TODO:  add validation step here that checks the model each rerender
     // this.set('isValidated', false)
     $('[data-toggle="tooltip"]').tooltip();
   },
+
   externalRepoMap: {},
   // externalSubmission: Ember.computed('metadataBlobNoKeys', function () { // eslint-disable-line
   //   if (this.get('metadataBlobNoKeys').Submission) {
@@ -21,35 +23,8 @@ export default Component.extend({
   metadata: Ember.computed('model.newSubmission.metadata', function () { // eslint-disable-line
     return JSON.parse(this.get('model.newSubmission.metadata'));
   }),
-  metadataBlobNoKeys: Ember.computed('model.newSubmission.metadata', function () { // eslint-disable-line
-    let metadataBlobNoKeys = [];
-    JSON.parse(this.get('model.newSubmission.metadata')).forEach((ele) => {
-      for (var key in ele.data) {
-        if (ele.data.hasOwnProperty(key)) {
-          let strippedData;
-          strippedData = ele.data[key];
-
-          if (key === 'authors') {
-            if (metadataBlobNoKeys['author(s)']) {
-              metadataBlobNoKeys['author(s)'] = _.uniqBy(metadataBlobNoKeys['author(s)'].concat(strippedData), 'author');
-            } else {
-              metadataBlobNoKeys['author(s)'] = strippedData;
-            }
-          } else if (key === 'container-title') {
-            metadataBlobNoKeys['journal-title'] = strippedData;
-          } else {
-            metadataBlobNoKeys[key] = strippedData;
-          }
-        }
-      }
-    });
-    for (var key in metadataBlobNoKeys) {
-      if (metadataBlobNoKeys.hasOwnProperty(key)) {
-        metadataBlobNoKeys[_.capitalize(key)] = metadataBlobNoKeys[key];
-        delete metadataBlobNoKeys[key];
-      }
-    }
-    return metadataBlobNoKeys;
+  metadataBlobNoKeys: Ember.computed('model.newSubmission.metadata', function () {
+    return this.get('metadataService').getDisplayBlob(this.get('model.newSubmission.metadata'));
   }),
   hasVisitedWeblink: Ember.computed('externalRepoMap', function () {
     return Object.values(this.get('externalRepoMap')).every(val => val === true);
