+++ app/controllers/submissions/detail.js
@@ -1,12 +1,14 @@
 import Controller from '@ember/controller';
-import _ from 'lodash';
 
 export default Controller.extend({
+  metadataService: Ember.inject.service('metadata-blob'),
+
   tooltips: function () {
     $(() => {
       $('[data-toggle="tooltip"]').tooltip();
     });
   }.on('init'),
+
   externalSubmission: Ember.computed('metadataBlobNoKeys', function () {
     return this.get('metadataBlobNoKeys').Submission;
   }),
@@ -70,31 +72,6 @@ export default Controller.extend({
     return null;
   }),
   metadataBlobNoKeys: Ember.computed('model.sub.metadata', function () {
-    let metadataBlobNoKeys = [];
-    JSON.parse(this.get('model.sub.metadata')).forEach((ele) => {
-      for (var key in ele.data) {
-        if (ele.data.hasOwnProperty(key)) {
-          let strippedData;
-          strippedData = ele.data[key];
-          if (key === 'authors') {
-            if (metadataBlobNoKeys['author(s)']) {
-              metadataBlobNoKeys['author(s)'] = _.uniqBy(metadataBlobNoKeys['author(s)'].concat(strippedData), 'author');
-            } else {
-              metadataBlobNoKeys['author(s)'] = strippedData;
-            }
-          } else {
-            metadataBlobNoKeys[key] = strippedData;
-          }
-        }
-      }
-    });
-
-    for (var key in metadataBlobNoKeys) {
-      if (metadataBlobNoKeys.hasOwnProperty(key)) {
-        metadataBlobNoKeys[_.capitalize(key)] = metadataBlobNoKeys[key];
-        delete metadataBlobNoKeys[key];
-      }
-    }
-    return metadataBlobNoKeys;
+    return this.get('metadataService').getDisplayBlob(this.get('model.sub.metadata'));
   }),
 });
